Made costDisplayable  protocol

+++ Health_Insurance_App/Plans/Plan Search/Models/CostDisplayable.swift
@@ -1,8 +1,16 @@
+//
+//  CostDisplayable.swift
+//  Health_Insurance_App
+//
+//  Created by Aaron Aslan on 9/23/26.
+//
+
+
 import Foundation
 
 protocol CostDisplayable {
     var id: String { get }
     var planName: String { get }
-    var deductible: Double { get }
-    var premium: Double { get }
+    var deductible: Decimal { get }
+    var premium: Decimal { get }
 }
